Basic verification
Should work, but differs from the Python version in the NTT operations. For some reason INTT(NTT(a) * NTT(b)) isn't what it should be. I've compared my implementation to some other I've found on the internet (also a Python version) and it should be right. I have no idea what the official one is doing