Combine signature salt and message, more fixes for hash_to_point
Previously verify module was tasked with combining signature salt and message when sending it to hash_to_point. Now this is delegated to the user/parent of verify. This makes the interface much simpler, since there are fewer signals. It shouldn't be that hard to use, since the first 5 blocks of data are always the salt and the remaining the message