Remove external integer multiplication from FLPMultiplier
The original authors had that there so they could use the same multipliers (hardware DSPs) for both NTT and FFT. We only need FTT here (I think), so we don't have to bother with the separate multipliers. Instead we just use normal multiplication and let Vivado handle it from there.